requirements.txt
python-dotenv
aiogram
loguru
